+++ frontend/src/features/auth/ui/index.ts
@@ -1,5 +1,5 @@
+export * from "./forms";
+
 export { AuthLayout } from "./AuthLayout";
 export { AuthProvider } from "./AuthProvider";
 export { LogoutButton } from "./LogoutButton";
-export { SignInForm } from "./SignInForm";
-export { SignUpForm } from "./SignUpForm";
